+++ example/lib/src/only_barrel_import.dart
@@ -1,0 +1,5 @@
+import 'package:example_core/util/example_core_util.dart';
+// Should be import 'package:example_core/example_core.dart';
+
+// ignore: unused_element
+final _ = ExampleCoreUtil();