+++ lib/src/analyzers/lint_analyzer/models/replacement.dart
@@ -6,9 +6,13 @@ class Replacement {
   /// The code with changes to replace original code with.
   final String replacement;
 
+  /// The priority of the change.
+  final int priority;
+
   /// Initialize a newly created [Replacement] with the given [comment] and [replacement].
   const Replacement({
     required this.comment,
     required this.replacement,
+    this.priority = 1,
   });
 }