+++ lib/src/analyzers/lint_analyzer/rules/rules_list/tag_name/tag_name_rule.dart
@@ -55,7 +55,7 @@ class TagNameRule extends DartRule {
                 source: source,
               ),
               message: _warning,
-              replacement: node.replacement,
+              replacements: [node.replacement],
             ))
         .toList(growable: false);
   }