+++ test/src/analyzers/lint_analyzer/reporters/reporters_list/json/lint_json_reporter_test.dart
@@ -73,10 +73,12 @@ void main() {
             'severity': 'warning',
             'message': 'simple message',
             'verboseMessage': 'verbose message',
-            'suggestion': {
-              'comment': 'replacement comment',
-              'replacement': 'function body',
-            },
+            'suggestions': [
+              {
+                'comment': 'replacement comment',
+                'replacement': 'function body',
+              }
+            ]
           },
         ]),
       );