+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.2.0
+- Add rule `use-design-system-items`.
+- Add rule `only-barrel-import`.
+- Allow to specify more than one suggestion for each rule.
+
 ## 3.1.1
 - Changed `prefer-media-query-direct-access` to `FlutterRule`.